shader_recompiler: Implement linear interpolation support
